TIP-860: move job parameters
