+++ UPGRADE.md
@@ -6,6 +6,14 @@ Several classes and services have been moved or renamed. The following commands 
 
 ```bash
 
+find ./src/ -type f -print0 | xargs -0 sed -i 's/Pim\\Bundle\\EnrichBundle\\Connector\\Job\\JobParameters\\ConstraintCollectionProvider\\ProductAndProductModelQuickExport/Akeneo\\Pim\\Enrichment\\Component\\Product\\Connector\\Job\\JobParameters\\ConstraintCollectionProvider\\ProductAndProductModelQuickExport/g'
+find ./src/ -type f -print0 | xargs -0 sed -i 's/Pim\\Bundle\\EnrichBundle\\Connector\\Job\\JobParameters\\ConstraintCollectionProvider\\ProductMassEdit/Akeneo\\Pim\\Enrichment\\Component\\Product\\Connector\\Job\\JobParameters\\ConstraintCollectionProvider\\ProductMassEdit/g'
+find ./src/ -type f -print0 | xargs -0 sed -i 's/Pim\\Bundle\\EnrichBundle\\Connector\\Job\\JobParameters\\ConstraintCollectionProvider\\ProductQuickExport/Akeneo\\Pim\\Enrichment\\Component\\Product\\Connector\\Job\\JobParameters\\ConstraintCollectionProvider\\ProductQuickExport/g'
+find ./src/ -type f -print0 | xargs -0 sed -i 's/Pim\\Bundle\\EnrichBundle\\Connector\\Job\\JobParameters\\ConstraintCollectionProvider\\SimpleMassEdit/Akeneo\\Pim\\Enrichment\\Component\\Product\\Connector\\Job\\JobParameters\\ConstraintCollectionProvider\\SimpleMassEdit/g'
+find ./src/ -type f -print0 | xargs -0 sed -i 's/Pim\\Bundle\\EnrichBundle\\Connector\\Job\\JobParameters\\DefaultValuesProvider\\ProductAndProductModelMassDelete/Akeneo\\Pim\\Enrichment\\Component\\Product\\Connector\\Job\\JobParameters\\DefaultValueProvider\\ProductAndProductModelMassDelete/g'
+find ./src/ -type f -print0 | xargs -0 sed -i 's/Pim\\Bundle\\EnrichBundle\\Connector\\Job\\JobParameters\\DefaultValuesProvider\\ProductMassEdit/Akeneo\\Pim\\Enrichment\\Component\\Product\\Connector\\Job\\JobParameters\\DefaultValueProvider\\ProductMassEdit/g'
+find ./src/ -type f -print0 | xargs -0 sed -i 's/Pim\\Bundle\\EnrichBundle\\Connector\\Job\\JobParameters\\DefaultValuesProvider\\ProductQuickExport/Akeneo\\Pim\\Enrichment\\Component\\Product\\Connector\\Job\\JobParameters\\DefaultValueProvider\\ProductQuickExport/g'
+find ./src/ -type f -print0 | xargs -0 sed -i 's/Pim\\Bundle\\EnrichBundle\\Connector\\Job\\JobParameters\\DefaultValuesProvider\\SimpleMassEdit/Akeneo\\Pim\\Enrichment\\Component\\Product\\Connector\\Job\\JobParameters\\DefaultValueProvider\\SimpleMassEdit/g'
 find ./src/ -type f -print0 | xargs -0 sed -i 's/Pim\\Bundle\\EnrichBundle\\Connector\\Processor\\AbstractProcessor/Akeneo\\Pim\\Enrichment\\Component\\Product\\Connector\\Processor\\MassEdit\\AbstractProcessor/g'
 find ./src/ -type f -print0 | xargs -0 sed -i 's/Pim\\Bundle\\EnrichBundle\\Connector\\Processor\\QuickExport\\ProductAndProductModelProcessor/Akeneo\\Pim\\Enrichment\\Component\\Product\\Connector\\Processor\\QuickExport\\ProductAndProductModelProcessor/g'
 find ./src/ -type f -print0 | xargs -0 sed -i 's/Pim\\Bundle\\EnrichBundle\\Connector\\Processor\\QuickExport\\ProductAndProductModelProcessor/Akeneo\\Pim\\Enrichment\\Component\\Product\\Connector\\Processor\\QuickExport\\ProductAndProductModelProcessor/g'
@@ -718,10 +726,10 @@ find ./src/ -type f -print0 | xargs -0 sed -i 's/Pim\\Component\\Connector\\Job\
 find ./src/ -type f -print0 | xargs -0 sed -i 's/Pim\\Component\\Connector\\Job\\JobParameters\\ConstraintCollectionProvider\\ProductModelCsvExport/Akeneo\\Pim\\Enrichment\\Component\\Product\\Connector\\Job\\JobParameters\\ConstraintCollectionProvider\\ProductModelCsvExport/g'
 find ./src/ -type f -print0 | xargs -0 sed -i 's/Pim\\Component\\Connector\\Job\\JobParameters\\ConstraintCollectionProvider\\ProductModelCsvImport/Akeneo\\Pim\\Enrichment\\Component\\Product\\Connector\\Job\\JobParameters\\ConstraintCollectionProvider\\ProductModelCsvImport/g'
 find ./src/ -type f -print0 | xargs -0 sed -i 's/Pim\\Component\\Connector\\Job\\JobParameters\\ConstraintCollectionProvider\\ProductXlsxExport/Akeneo\\Pim\\Enrichment\\Component\\Product\\Connector\\Job\\JobParameters\\ConstraintCollectionProvider\\ProductXlsxExport/g'
-find ./src/ -type f -print0 | xargs -0 sed -i 's/Pim\\Component\\Connector\\Job\\JobParameters\\DefaultValuesProvider\\ProductCsvImport/Akeneo\\Pim\\Enrichment\\Component\\Product\\Connector\\Job\\JobParameters\\DefaultValuesProvider\\ProductCsvImport/g'
-find ./src/ -type f -print0 | xargs -0 sed -i 's/Pim\\Component\\Connector\\Job\\JobParameters\\DefaultValuesProvider\\ProductModelCsvExport/Akeneo\\Pim\\Enrichment\\Component\\Product\\Connector\\Job\\JobParameters\\DefaultValuesProvider\\ProductModelCsvExport/g'
-find ./src/ -type f -print0 | xargs -0 sed -i 's/Pim\\Component\\Connector\\Job\\JobParameters\\DefaultValuesProvider\\ProductModelCsvImport/Akeneo\\Pim\\Enrichment\\Component\\Product\\Connector\\Job\\JobParameters\\DefaultValuesProvider\\ProductModelCsvImport/g'
-find ./src/ -type f -print0 | xargs -0 sed -i 's/Pim\\Component\\Connector\\Job\\JobParameters\\DefaultValuesProvider\\ProductXlsxExport/Akeneo\\Pim\\Enrichment\\Component\\Product\\Connector\\Job\\JobParameters\\DefaultValuesProvider\\ProductXlsxExport/g'
+find ./src/ -type f -print0 | xargs -0 sed -i 's/Pim\\Component\\Connector\\Job\\JobParameters\\DefaultValuesProvider\\ProductCsvImport/Akeneo\\Pim\\Enrichment\\Component\\Product\\Connector\\Job\\JobParameters\\DefaultValueProvider\\ProductCsvImport/g'
+find ./src/ -type f -print0 | xargs -0 sed -i 's/Pim\\Component\\Connector\\Job\\JobParameters\\DefaultValuesProvider\\ProductModelCsvExport/Akeneo\\Pim\\Enrichment\\Component\\Product\\Connector\\Job\\JobParameters\\DefaultValueProvider\\ProductModelCsvExport/g'
+find ./src/ -type f -print0 | xargs -0 sed -i 's/Pim\\Component\\Connector\\Job\\JobParameters\\DefaultValuesProvider\\ProductModelCsvImport/Akeneo\\Pim\\Enrichment\\Component\\Product\\Connector\\Job\\JobParameters\\DefaultValueProvider\\ProductModelCsvImport/g'
+find ./src/ -type f -print0 | xargs -0 sed -i 's/Pim\\Component\\Connector\\Job\\JobParameters\\DefaultValuesProvider\\ProductXlsxExport/Akeneo\\Pim\\Enrichment\\Component\\Product\\Connector\\Job\\JobParameters\\DefaultValueProvider\\ProductXlsxExport/g'
 find ./src/ -type f -print0 | xargs -0 sed -i 's/Pim\\Component\\Connector\\Processor\\Denormalization\\Product\\FindProductToImport/Akeneo\\Pim\\Enrichment\\Component\\Product\\Connector\\Processor\\Denormalizer\\FindProductToImport/g'
 find ./src/ -type f -print0 | xargs -0 sed -i 's/Pim\\Component\\Connector\\Processor\\Denormalization\\ProductAssociationProcessor/Akeneo\\Pim\\Enrichment\\Component\\Product\\Connector\\Processor\\Denormalizer\\ProductAssociationProcessor/g'
 find ./src/ -type f -print0 | xargs -0 sed -i 's/Pim\\Component\\Connector\\Processor\\Denormalization\\ProductModelAssociationProcessor/Akeneo\\Pim\\Enrichment\\Component\\Product\\Connector\\Processor\\Denormalizer\\ProductModelAssociationProcessor/g'
