+++ src/Pim/Bundle/EnrichBundle/Resources/config/job_constraints.yml
@@ -1,8 +1,8 @@
 parameters:
-    pim_enrich.connector.job.job_parameters.constraint_collection_provider.product_mass_edit.class: Pim\Bundle\EnrichBundle\Connector\Job\JobParameters\ConstraintCollectionProvider\ProductMassEdit
-    pim_enrich.connector.job.job_parameters.constraint_collection_provider.product_and_product_model_quick_export.class: Pim\Bundle\EnrichBundle\Connector\Job\JobParameters\ConstraintCollectionProvider\ProductAndProductModelQuickExport
-    pim_enrich.connector.job.job_parameters.constraint_collection_provider.product_quick_export.class: Pim\Bundle\EnrichBundle\Connector\Job\JobParameters\ConstraintCollectionProvider\ProductQuickExport
-    pim_enrich.connector.job.job_parameters.constraint_collection_provider.simple_mass_edit.class: Pim\Bundle\EnrichBundle\Connector\Job\JobParameters\ConstraintCollectionProvider\SimpleMassEdit
+    pim_enrich.connector.job.job_parameters.constraint_collection_provider.product_mass_edit.class: Akeneo\Pim\Enrichment\Component\Product\Connector\Job\JobParameters\ConstraintCollectionProvider\ProductMassEdit
+    pim_enrich.connector.job.job_parameters.constraint_collection_provider.product_and_product_model_quick_export.class: Akeneo\Pim\Enrichment\Component\Product\Connector\Job\JobParameters\ConstraintCollectionProvider\ProductAndProductModelQuickExport
+    pim_enrich.connector.job.job_parameters.constraint_collection_provider.product_quick_export.class: Akeneo\Pim\Enrichment\Component\Product\Connector\Job\JobParameters\ConstraintCollectionProvider\ProductQuickExport
+    pim_enrich.connector.job.job_parameters.constraint_collection_provider.simple_mass_edit.class: Akeneo\Pim\Enrichment\Component\Product\Connector\Job\JobParameters\ConstraintCollectionProvider\SimpleMassEdit
 
 services:
     pim_enrich.connector.job.job_parameters.constraint_collection_provider.product_mass_edit:
