+++ src/Pim/Bundle/EnrichBundle/spec/Connector/Job/JobParameters/ConstraintCollectionProvider/ProductAndProductModelQuickExportSpec.php
@@ -5,7 +5,7 @@ namespace spec\Pim\Bundle\EnrichBundle\Connector\Job\JobParameters\ConstraintCol
 use Akeneo\Tool\Component\Batch\Job\JobInterface;
 use Akeneo\Tool\Component\Batch\Job\JobParameters\ConstraintCollectionProviderInterface;
 use PhpSpec\ObjectBehavior;
-use Pim\Bundle\EnrichBundle\Connector\Job\JobParameters\ConstraintCollectionProvider\ProductAndProductModelQuickExport;
+use Akeneo\Pim\Enrichment\Component\Product\Connector\Job\JobParameters\ConstraintCollectionProvider\ProductAndProductModelQuickExport;
 use Symfony\Component\Validator\Constraints\Collection;
 use Symfony\Component\Validator\Constraints\NotBlank;
